# docs/tings-resale-tags-api-docs.json
{"openapi":"3.0.3","info":{"version":"9.5.2","title":"@fastify/swagger"},"components":{"securitySchemes":{"apiKey":{"type":"apiKey","in":"header","name":"api-key"}},"schemas":{}},"paths":{"/get-tags-autocomplete":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","required":["query"],"properties":{"category":{"type":"array","items":{"type":"string"},"example":[]},"categoryExclude":{"type":"array","items":{"type":"string"},"example":[]},"brand":{"type":"array","items":{"type":"string"},"example":[]},"brandExclude":{"type":"array","items":{"type":"string"},"example":[]},"tag":{"type":"array","items":{"type":"string"},"example":[]},"tagExclude":{"type":"array","items":{"type":"string"},"example":[]},"query":{"type":"string"},"limit":{"type":"number","default":20}}}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/get-listings":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"category":{"type":"array","items":{"type":"string"},"example":[]},"categoryExclude":{"type":"array","items":{"type":"string"},"example":[]},"brand":{"type":"array","items":{"type":"string"},"example":[]},"brandExclude":{"type":"array","items":{"type":"string"},"example":[]},"tag":{"type":"array","items":{"type":"string"},"example":[]},"tagExclude":{"type":"array","items":{"type":"string"},"example":[]},"query":{"type":"string"},"usePostgres":{"type":"boolean","default":true},"listingsPerTagLimit":{"type":"number","default":1000},"tagsLimit":{"type":"number","default":100},"listingsSampleLimit":{"type":"number","default":100}}}}}},"responses":{"200":{"description":"Default Response"}}}},"/get-category-autocomplete":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","required":["query"],"properties":{"query":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/get-brands-autocomplete":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","required":["query"],"properties":{"category":{"type":"array","items":{"type":"string"}},"query":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/get-estimate-for-product":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","required":["name"],"properties":{"name":{"type":"string"},"brand":{"type":"string"},"category":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/get-estimate-new":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","required":["searchTerm"],"properties":{"searchTerm":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"Default Response"}}}}}}